Module Brzo.Cli

Cli interaction.

Man pages fragments

val s_domains : string

s_domains is a section that lists the domains.

val s_outcomes : string

s_outcomes is a manual section for outcome options.

val s_outcome_mode : string

s_outcome_opts is a manual section for outcome specific options.

val s_outcome_opts : string

s_outcome_opts is a manual section for outcome specific options.

val man_see_manual : Cmdliner.Manpage.block

man_see_manual is a cmdliner manual fragment enticing to read the manual.

Cli configuration

val conf : auto_cwd_root:bool -> use_brzo_file:bool -> domain:Pre_domain.t option -> all_domains:Pre_domain.t list -> Conf.t Cmdliner.Term.t

conf ~use_brzo_file ~domain ~all_domain is a Cmdliner term for configuration with:

auto_cwd_root, allows the root to fall back to Conf.cwd if the root can't be determined. This allows commands that work e.g. on the _b0 directory to work after the BRZO file marking the root might have been deleted.

use_brzo_file, if true the BRZO file is consulted.

domain, if Some d the cli has the arguments for the pre-domain d.

all_domains is the list of domains knows to brzo as pre-domains.

Term evaluation performs a few side effects. In particular it setups colored output and logging via B0_std_ui.setup.