Module Brzo.Exit

Program exits.

val conf_error : B00_std.Os.Exit.t

conf_error indicates a brzo configuration error (e.g. in a BRZO file.

val no_build_outcome : B00_std.Os.Exit.t

no_build_outcome indicates there's no build outcome on operations that request one.

val no_such_sexp_path : B00_std.Os.Exit.t

no_such_sexp_path indicates that a specified s-expression path does not exist.

val ok : B00_std.Os.Exit.t

ok is the zero exit code.

val outcome_build_error : B00_std.Os.Exit.t

outcome_build_error indicates an outcome build error.

val outcome_action_error : B00_std.Os.Exit.t

outcome_action_error indicates an outcome action error.

val some_error : B00_std.Os.Exit.t

some_error indicates an indiscrimante error reported on stderr.

val undefined_domain : B00_std.Os.Exit.t

undefined_domain indicates use of an undefined domain in BRZO file.

module Info : sig ... end

Cmdliner documentation.