Module Brzo.Memo

Memo helpers.

val copy_file : B00.Memo.t -> src_root:B00_std.Fpath.t -> dst_root:B00_std.Fpath.t -> B00_std.Fpath.t -> unit

copy m ~src_root ~dst_root src copies src to destination root_path ~src_root ~dst_root src.

val ensure_exec_build : B00.Memo.t -> srcs:B00_fexts.map -> need_ext:B00_std.Fpath.ext -> unit B00_std.Fut.t

ensure_exe_build m ~srcs ~need_ext k continues iff there's at least a file in srcs that has exention need_ext and fails with a message that no executable can be built otherwise.

val run : with_log:B00_std.Fpath.t option -> B00.Memo.t -> (unit -> 'a B00_std.Fut.t) -> ('a, unit) result