Module Brzo_b0_c.Inc_deps

C include dependencies (via gcc -M)

Parses the format output by -M which is basically a Makefile rule.

val write : ?deps:B00_std.Fpath.t list -> B00.Memo.t -> src:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

write m ~deps src o writes dependencies of src in file o. deps indicates these files should become ready first: we do not use -MG as it leads to output race conditions so one needs to make sure all generated includes that may end up being used src are generated before asking for dependencies.

val read : B00.Memo.t -> src:B00_std.Fpath.t -> B00_std.Fpath.t -> B00_std.Fpath.t list B00_std.Fut.t

read m file reads dependencies produced by write in file for source src (whose parent directory is used to make paths absolute, see of_string).

Low-level parsing

val of_string : ?file:B00_std.Fpath.t -> root:B00_std.Fpath.t -> string -> (B00_std.Fpath.t list, string) result

of_string ~file ~src s parses -M dependencies data from s assuming it was read from file file (for error reporting, defaults to Os.File.dash). root is used to make file paths absolute (whether file paths end up being absolute or not seem to depend on the cwd when write is invoked.