Module Brzo_b0_js_of_ocaml.Js_of_ocaml

js_of_ocaml compilation.

Js_of_ocaml

val compile : ?args:B00_std.Cmd.t -> B00.Memo.t -> byte_exe:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

compile m ~byte_exe ~o compiles the byte code executable byte_exe to the JavaScript file o.

val compile_toplevel : ?args:B00_std.Cmd.t -> B00.Memo.t -> byte_exe:B00_std.Fpath.t -> mod_names:B00_std.Fpath.t -> o:B00_std.Fpath.t -> unit

compile m ~byte_exe ~mod_names ~o compiles the byte code executable byte_exe to a toplevel in which the modules mentioned in mod_names are made visible.

val link : ?args:B00_std.Cmd.t -> B00.Memo.t -> jss:B00_std.Fpath.t list -> o:B00_std.Fpath.t -> unit

link m ~jss ~o links the JavaScript files jss into o.

val tty_glue : exe:B00_std.Fpath.t -> cwd:B00_std.Fpath.t -> env:B00_std.Os.Env.t -> args:string list -> string

val write_page : ?toplevel_css:bool -> ?generator:string -> ?lang:string -> ?scripts:string list -> ?styles:string list -> ?title:string -> B00.Memo.t -> o:B00_std.Fpath.t -> unit

write_page m ~title ~o writes to file o a full HTML document whose body contains only a B0_web.Htmlg.noscript element that entices the user, in english, to enable javascript. title defaults to the basename of o without its extension, for the other arguments and more information see B0_web.Htmlg.basic_page.

toplevel_css if true adds a bit of styling for an #ocaml element that is supposed to hold a toplevel session (defaults to false).

val toplevel_ui_src : string

toplevel_ui_src a simple toplevel user interface written in OCaml. That can be compiled and linked when using compile_toplevel to get a simple prompt.