Module Brzo_b0_latex.Bibdoi

Lists of DOIs and their resolution.

.bibdoi files

type t

The type for .bibdoi files.

val sexp : t -> B00_serialk_sexp.Sexp.t

sexp b is b's underlying s-expression.

val dois : t -> Doi.t list

dois b is the list of DOIs in b.

val of_string : ?file:B00_std.Fpath.t -> string -> (t, string) result

of_string s parses a .bibdoi file from s assuming it was read from file file (for error reporting, defaults to Os.File.dash.

val to_bibtex : ?resolver:string -> B00_http.Httpr.t -> t -> (string, string) result

to_bibtex r ~resolver b resolves the DOIs of b to bibtex entries using resolver (defaults to Doi.default_resolver.