Module Brzo_b0_latex.Doi

DOIs and their resolution.

val default_resolver : B00_http.Uri.t

default_resolver is the default resolver used to resolve DOIs. This is https://doi.org.

type t = string

The type for DOIs.

val pp : t B00_std.Fmt.t

pp formats a DOI.

val default_bib_format : string

default_bib_format is the default bibliographic format. This is "application/x-bibtex; charset=utf-8"

val oneline_bib_format : string

oneline_bib_format is "text/bibliography; charset=utf-8".

val resolve_to_uri : ?resolver:string -> B00_http.Httpr.t -> t -> (string, string) result

resolve_to_uri r ~resolver doi resolve doi with resolver to an URI with resolver (defaults to default_resolver).

val resolve_to_bib : ?format:string -> ?resolver:B00_http.Uri.t -> B00_http.Httpr.t -> t -> (string, string) result

resolve_to_bib r ~resolver doi resolves doi to a bibliographic entry with resolver (defaults to default_resolver).