Module Brzo_b0_latex.Fls

.fls files reader.

A .fls is generated by TeX tools whenever the -record option is specified. It indicates all the files read and written by the LaTeX generation.

.fls Files

type t

The type for contents of .fls files.

val reads : t -> B00_std.Fpath.Set.t

reads f are absolute paths to file reads.

val writes : t -> B00_std.Fpath.Set.t

writes f are absolute paths to file writes.

val of_string : ?file:B00_std.Fpath.t -> string -> (t, string) result

of_string s parses .fls data from s assuming it was read from file file (for error reporting, defaults to Os.File.dash).