Module Brzo_domain

Brzo domains.

Domains

module type T = sig ... end

Signature implemented by domains

type t =

| V : (module T with type Conf.t = 'a) -> t	(*
The type for domains.
*)

val name : t -> string

name d is the identifier of domain d.

val fingerprint : t -> B00_fexts.t

fingerprint d is the fingerprint of d.

val pre_domain : t -> Brzo.Pre_domain.t

pre_domain d is the pre-domain of d.

val run : Brzo.Conf.t -> t -> B00_std.Os.Exit.t

run c d runs domain d with configuration c.

Finding domains

val find : string -> t list -> (t, string) result

find name ds is the domain identified named n in ds.

val of_conf : Brzo.Conf.t -> t list -> (t, string) result

of_conf c ds is the domain from ds selected in configuration c.