Module Brzo_latex.Conf

Domain configuration.

Domain configuration

type t

The type for domain specific configurations.

val tid : t B00_std.Tid.t

tid is a type identifier for the configuration.

val keys : B00_std.String.Set.t

keys is the set of keys that are allowed to be part of the domain specific BRZO file dictionary.

val parse : (B00_serialk_sexp.Sexp.t * B00_serialk_sexp.Sexpq.path) -> (t, string) result

parse is a function that parses the BRZO file domain specific dictionary into a configuration.

val parse_with_cli : ((B00_serialk_sexp.Sexp.t * B00_serialk_sexp.Sexpq.path) -> (t, string) result) Cmdliner.Term.t

parse_with_cli is a cmdliner term that has the command line interface for the configuration and a function that given the configuration s-expression dictionary for the domain should derive the configuration appropriately (basically it should do the job of merging cli and file specified configuration desires).

val pp : t B00_std.Fmt.t

pp formats configurations.