Module Brzo_ocaml_cmi

Compiled interfaces.

type t

The type for compiled interfaces.

val read : B00.Memo.t -> B00_std.Fpath.t -> t B00_std.Fut.t

read f reads an object file from f.

val file : t -> B00_std.Fpath.t

file cmi is the file path of cmi.

val mod_ref : t -> B00_ocaml.Mod.Ref.t

mod_ref cmi is the module reference of cmi.

val deps : t -> B00_ocaml.Mod.Ref.Set.t

deps cmi is the set of modules interfaces imported by cmi.

val mod_names : t -> B00_ocaml.Mod.Name.Set.t

mod_names cmi are the unqualified module names defined by cmi (including its name). Sligthly wrong because stops at module aliases, these are not resolved to further cmis.

val pp : t B00_std.Fmt.t

pp formats a compiled interface.