Module Brzo_read_cmi

val read : B00_std.Fpath.t -> (string * Digest.t * B00_std.String.Set.t * (string * Digest.t) list, string) result

read cmi is Ok (name, digest, names, deps) with

name the module name of the interface.

digest the digest of the inteface.

names are the unqualified module names defined in the interface, including name itself. This is slightly incomplete because it stops at module aliases: these would need to be resolved to further cmis to find the names therein.

deps are the digested module interfaces imported by this interface.