Module Brzo_tie_conf

Ties the configuration knot.

Configuration in Brzo.Conf needs the Brzo_domain_list which uses Brzo_domain which uses Brzo.Conf. This module allows to break the cyclic dependency by transforming Brzo_domains to Brzo.Pre_domains.

Configuration

val for_domain : domain:Brzo_domain.t -> Brzo.Conf.t Cmdliner.Term.t

conf_for_domain ~domain is a cmdliner term for configuration for domain domain. The root is determined and the BRZO file is consulted.

val use_brzo_file : Brzo.Conf.t Cmdliner.Term.t

use_brzo_file is a cmdliner term for brzo configuration when no domain outcome is involved. The root is determined and the BRZO file is consulted.

val no_brzo_file : Brzo.Conf.t Cmdliner.Term.t

no_brzo_file is a cmdliner term for brzo configuration when no domain outcome is involved. The root is determined but the BRZO file is not consulted.

val auto_cwd_root_and_no_brzo_file : Brzo.Conf.t Cmdliner.Term.t

auto_root_root_no_brzo_file is a cmdliner term for brzo configuration when no domain outcome is involved. The root falls back to cwd if it can be determined and the BRZO file is not consulted.